FIxes + user registration screen

+++ src/components/screens/settings-screen/settings-screen.tsx
@@ -76,8 +76,8 @@ class Settings extends React.Component<Props, State> {
     });
     try {
       const userSettingsApi = Api.getUsersApi(accessToken);
-      const userSettings = await userSettingsApi.getUserSettings();
-      const { homeAddress } = userSettings;
+      const user = await userSettingsApi.getUser({userId: accessToken?.userId || ""});
+      const { homeAddress } = user;
 
       if (homeAddress) {
         this.setState({ 
@@ -284,7 +284,7 @@ class Settings extends React.Component<Props, State> {
     }
 
     const usersApi = Api.getUsersApi(accessToken);
-    await usersApi.deleteUser();
+    await usersApi.deleteUser({ userId: accessToken.userId || "" });
     window.location.href = "/";
   }
 
@@ -379,19 +379,22 @@ class Settings extends React.Component<Props, State> {
     const userSettingsApi = Api.getUsersApi(accessToken);
     if (streetAddress === "" && postalCode === "" && city === "" && country === "") {
       try {
-        const userSettings = {
-          showMobileWelcomeScreen: false,
-          pollutantPenalties: {},
-          pollutantThresholds: {}
-        };
+        const existingUser = await userSettingsApi.getUser({userId: accessToken.userId || ""});
         if (userSettingsExist) {
-          await userSettingsApi.updateUserSettings({ 
-            userSettings: userSettings
+          await userSettingsApi.updateUser({ 
+            user: {
+              ...existingUser,
+              showMobileWelcomeScreen: false,
+              pollutantPenalties: []
+            },
+            userId: accessToken.userId || ""
           });
         } else {
-          await userSettingsApi.createUserSettings({ 
-            userSettings: userSettings
-          });
+          //TODO: check if creating new User is needed here
+          return;
+          // await userSettingsApi.createUserSettings({ 
+          //   userSettings: userSettings
+          // });
         }
       } catch (error) {
         console.log(error);
@@ -413,13 +416,22 @@ class Settings extends React.Component<Props, State> {
         };
 
         if (userSettingsExist) {
-          await userSettingsApi.updateUserSettings({
-            userSettings: userSettings
+          const existingUser = await userSettingsApi.getUser({userId: accessToken.userId || ""});
+          await userSettingsApi.updateUser({
+            user: {
+              ...existingUser,
+              showMobileWelcomeScreen: false,
+              pollutantPenalties: [],
+              homeAddress: homeAddress
+            },
+            userId: accessToken.userId || ""
           });
         } else {
-          await userSettingsApi.createUserSettings({
-            userSettings: userSettings
-          });
+          //TODO: check if creating new User is needed here
+          return;
+          // await userSettingsApi.createUserSettings({
+          //   userSettings: userSettings
+          // });
         }
       } catch (error) {
         console.log(error);
